--- server/.air.toml
@@ -1,4 +1,0 @@
-[build]
-cmd = "go build -o tmp/main.exe ./cmd/main.go && tmp/main.exe"
-bin = "tmp/main.exe"
-exclude_dir = ["tmp", "vendor"]
