+++ server/internal/api/portfolio_service.go
@@ -3,7 +3,6 @@ package api
 import (
 	"encoding/json"
 	"fmt"
-	"log"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -24,89 +23,6 @@ func (s *portfolioService) RegisterRoutes(r *gin.Engine) {
 	r.GET("/projects", s.GetProject)
 	r.GET("/connect_option", s.GetConnectOption)
 	r.GET("/user_info", s.GetUserInfo)
-}
-
-func (s *portfolioService) GetProject(ctx *gin.Context) {
-
-	data, err := getLocalJson("projects.json")
-	if err != nil {
-		models.WriteError(ctx.Writer, *err)
-		return
-	}
-
-	var jsonData map[string]interface{}
-
-	if err := json.Unmarshal([]byte(data), &jsonData); err != nil {
-		models.WriteError(ctx.Writer, models.APIError{
-			Code:    strconv.Itoa(http.StatusInternalServerError),
-			Message: fmt.Sprintf("error unmarshaling JSON: %v", err),
-		})
-		return
-	}
-
-	projectsData, ok := jsonData["data"].([]interface{})
-	if !ok {
-		models.WriteError(ctx.Writer, models.APIError{
-			Code:    strconv.Itoa(http.StatusInternalServerError),
-			Message: "invalid structure for 'data' in JSON",
-		})
-		return
-	}
-
-	var projects []models.Project
-	for _, p := range projectsData {
-		projectMap, ok := p.(map[string]interface{})
-		if !ok {
-			models.WriteError(ctx.Writer, models.APIError{
-				Code:    strconv.Itoa(http.StatusInternalServerError),
-				Message: "invalid project structure in JSON",
-			})
-			return
-		}
-
-		projectJSON, err := json.Marshal(projectMap)
-		if err != nil {
-			models.WriteError(ctx.Writer, models.APIError{
-				Code:    strconv.Itoa(http.StatusInternalServerError),
-				Message: fmt.Sprintf("error marshaling project: %v", err),
-			})
-			return
-		}
-
-		var project models.Project
-		if err := json.Unmarshal(projectJSON, &project); err != nil {
-			models.WriteError(ctx.Writer, models.APIError{
-				Code:    strconv.Itoa(http.StatusInternalServerError),
-				Message: fmt.Sprintf("error unmarshaling project: %v", err),
-			})
-			return
-		}
-
-		if project.Thumbnail != nil {
-			m := fullPath(*project.Thumbnail)
-			project.Thumbnail = &m
-		}
-
-		projects = append(projects, project)
-	}
-
-	models.WriteOk(ctx.Writer, projects)
-}
-
-var basDir *string
-
-func fullPath(m models.Media) models.Media {
-	if basDir == nil {
-		dir, err := os.Getwd()
-		if err != nil {
-			log.Fatalf("failed to get OS path %v", err)
-			return m
-		}
-
-		basDir = &dir
-	}
-	m.Url = filepath.Join(*basDir, "database", "images", m.Url)
-	return m
 }
 
 func (s *portfolioService) GetConnectOption(ctx *gin.Context) {
@@ -134,7 +50,7 @@ func (s *portfolioService) getJsonResponse(ctx *gin.Context, fileName string) {
 		return
 	}
 
-	models.WriteOk(ctx.Writer, jsonData)
+	models.WriteJson(ctx.Writer, http.StatusAccepted, jsonData)
 }
 
 // GetLocalJson reads a JSON file from /database/json/ and returns its content as a string.
